Refactor game setup components
