+++ Assets/Scripts/Events/GameEvents.cs
@@ -1,48 +1,37 @@
 using System;
 using System.Collections.Generic;
-using TicTacToe.GameProgression;
 using UnityEngine;
 
-namespace TicTacToe.GameManagement
+public class GameEvents : MonoBehaviour
 {
-    public class GameEvents : MonoBehaviour
+    public static GameEvents Instance;
+
+    void Awake()
     {
-        public static GameEvents Instance;
+        Instance = this;
+    }
 
-        void Awake()
-        {
-            Instance = this;
-        }
+    public event Action onGameStart;
+    public void GameStart()
+    {
+        onGameStart?.Invoke();
+    }
 
-        public event Action onGameStart;
-        public void GameStart()
-        {
-            onGameStart?.Invoke();
-        }
+    public event Action<enGameState> onGameEnded;
+    public void GameEnded(enGameState gameResult)
+    {
+        onGameEnded?.Invoke(gameResult);
+    }
 
-        public event Action onGameRestart;
-        public void GameRestart()
-        {
-            onGameRestart?.Invoke();
-        }
+    public event Action<Vector2Int, enSymbol> onMadeMove;
+    public void MadeMove(Vector2Int position, enSymbol symbol)
+    {
+        onMadeMove?.Invoke(position, symbol);
+    }
 
-        public event Action<enGameState> onGameEnded;
-        public void GameEnded(enGameState gameResult)
-        {
-            onGameEnded?.Invoke(gameResult);
-        }
-
-        public event Action<Vector2Int, enSymbol> onMadeMove;
-        public void MadeMove(Vector2Int position, enSymbol symbol)
-        {
-            onMadeMove?.Invoke(position, symbol);
-        }
-
-        public event Action<List<Vector2Int>> onUndoLastMove;
-        public void UndoLastMove(List<Vector2Int> lastMove)
-        {
-            onUndoLastMove?.Invoke(lastMove);
-        }
+    public event Action<List<Vector2Int>> onUndoLastMove;
+    public void UndoLastMove(List<Vector2Int> lastMove)
+    {
+        onUndoLastMove?.Invoke(lastMove);
     }
 }
-
